refactor(client)!: NEW Client class (Protocol-free) composes LegacyClient
`Client` no longer extends `Protocol`. It owns the 2026 stateless send
path (_isStateless/_buildMeta/_withMeta/_collect/_send/_negotiate/
subscribe/_listChangedLoop) and the typed request methods
(callTool/listTools/getPrompt/listPrompts/readResource/listResources/
listResourceTemplates/complete/setLoggingLevel), lifted verbatim from
the sectioned `LegacyClient` with field reads re-pointed at
`this.config.*` / `this._legacy.*`.

`LegacyClient` keeps the session-dependent block. Both share one handler
registry via `_legacy._dispatch`/`_legacy.setRequestHandler`; negotiated
server state lives on `_legacy` (single source of truth) and is written
by both `_negotiate` (via `_setNegotiated`) and `_initialize`.
`client.legacy` is the explicit escape hatch.

LegacyClient body changes are limited to:
- `_assertSession()` guard inserted at top of ping / subscribeResource /
  unsubscribeResource / sendRootsListChanged
- ctor: dropped now-unused `_enforceStrictCapabilities` /
  `_pendingListChangedConfig` / `_cachedToolOutputValidators` fields
  (lifted to NEW Client)
- `_initialize` / `_setupListChangedHandler` visibility widened to @internal
