+++ examples/client/src/elicitationUrlExample.ts
@@ -663,7 +663,7 @@ async function listTools(): Promise<void> {
             method: 'tools/list',
             params: {}
         };
-        const toolsResult = await client.request(toolsRequest);
+        const toolsResult = await client.legacy.request(toolsRequest);
 
         console.log('Available tools:');
         if (toolsResult.tools.length === 0) {
