+++ packages/client/src/client/client.examples.ts
@@ -9,7 +9,7 @@
 
 import type { Prompt, Resource, Tool } from '@modelcontextprotocol/core';
 
-import { LegacyClient as Client } from './legacyClient.js';
+import { Client } from './client.js';
 import { SSEClientTransport } from './sse.js';
 import { StdioClientTransport } from './stdio.js';
 import { StreamableHTTPClientTransport } from './streamableHttp.js';
