+++ packages/client/src/index.ts
@@ -52,10 +52,11 @@ export {
     PrivateKeyJwtProvider,
     StaticPrivateKeyJwtProvider
 } from './client/authExtensions.js';
-export type { ClientOptions } from './client/legacyClient.js';
-export { getSupportedElicitationModes, LegacyClient, LegacyClient as Client } from './client/legacyClient.js';
+export type { ClientOptions } from './client/client.js';
+export { Client, getSupportedElicitationModes } from './client/client.js';
 export type { DiscoverAndRequestJwtAuthGrantOptions, JwtAuthGrantResult, RequestJwtAuthGrantOptions } from './client/crossAppAccess.js';
 export { discoverAndRequestJwtAuthGrant, exchangeJwtAuthGrant, requestJwtAuthorizationGrant } from './client/crossAppAccess.js';
+export { LegacyClient } from './client/legacyClient.js';
 export type { LoggingOptions, Middleware, RequestLogger } from './client/middleware.js';
 export { applyMiddlewares, createMiddleware, withLogging, withOAuth } from './client/middleware.js';
 export type { SSEClientTransportOptions } from './client/sse.js';
